requirements.txt
sqlalchemy
pydantic
psycopg2
python-jose